Added a side panel featuring playlists
Created a playlists table in the database to keep track of songs. Also
renamed the table ‘Library’ to ‘Songs’ in the database to more
accurately represent the Song object. Created an association table
between playlists and songs because it is a many-to-many relationship.
Each playlist can contain 0..* songs, and each song can be contained by
1..* playlist (the 1 meaning the Library). All of these changes are in
the ‘Updates to database.sql’ file and should be run once in Netbeans.

Cleaned up the DataBaseModel class methods to work with prepared
statements instead of regular statements. Also simplified insert
methods. Added a method to insert a playlist into the database, and
retrieve all playlist names from the database. Insert song method has a
new parameter called playlistName. If this is null, the method will
insert the song info into the Songs table. Otherwise, it will add the
song path to the association table, essentially adding the song to a
playlist, not the Library. This allows for a simple link between the
song already in the library, and the playlist. It also allows for
duplicate songs to show up in any playlist.

Created a side panel in the View class that displays the library node
which is the main song table, as well as the Playlists node which
contains all playlists. Added a menu option to create a playlist and
the side panel updates automatically each time a playlist is created.
Each playlist name must be unique.